Create separate audio service
